Initial commit for a family feud style game for educators.
We now have a Title Menu that shows a logo and some text and then advances to the Team Setup Menu on key press / mouse click.

Also setup an array to hold entities on each screen to make it a bit easier when removing objects between each state.

Getting a basic foundation in-place and figuring out some haxe/luxeengine things.

This builds on some of the initial examples / tutorials from the luxengine site:
http://luxeengine.com/docs/guide.html#gettingstarted
https://github.com/underscorediscovery/luxe/tree/master/samples/alphas/2_0_0010 (for state setup)